requirements.txt
google-adk
google-generativeai
streamlit
pandas
google-cloud-bigquery
python-dotenv
